Moved FailAlertMessage.js into the components folder.
